+++ demo/app/demo-custom-items/index.js
@@ -1,17 +1,13 @@
 import React, { Component } from 'react'
-import moment from 'moment'
 
 import Timeline from 'react-calendar-timeline'
 // import containerResizeDetector from 'react-calendar-timeline/lib/resize-detector/container'
 
 import generateFakeData from '../generate-fake-data'
+import { add, formatISO, startOfDay } from 'date-fns'
 
-var minTime = moment()
-  .add(-6, 'months')
-  .valueOf()
-var maxTime = moment()
-  .add(6, 'months')
-  .valueOf()
+var minTime = add(new Date(), {months: -6}).valueOf()
+var maxTime = add(new Date(), {months: 6}).valueOf()
 
 var keys = {
   groupIdKey: 'id',
@@ -30,13 +26,8 @@ export default class App extends Component {
     super(props)
 
     const { groups, items } = generateFakeData()
-    const defaultTimeStart = moment()
-      .startOf('day')
-      .toDate()
-    const defaultTimeEnd = moment()
-      .startOf('day')
-      .add(1, 'day')
-      .toDate()
+    const defaultTimeStart = startOfDay(new Date())
+    const defaultTimeEnd = add(startOfDay(new Date()), {days: 1})
 
     this.state = {
       groups,
@@ -47,27 +38,27 @@ export default class App extends Component {
   }
 
   handleCanvasClick = (groupId, time, event) => {
-    console.log('Canvas clicked', groupId, moment(time).format())
+    console.log('Canvas clicked', groupId, formatISO(time))
   }
 
   handleCanvasContextMenu = (group, time, e) => {
-    console.log('Canvas context menu', group, moment(time).format())
+    console.log('Canvas context menu', group, formatISO(time))
   }
 
   handleItemClick = (itemId, _, time) => {
-    console.log('Clicked: ' + itemId, moment(time).format())
+    console.log('Clicked: ' + itemId, formatISO(time))
   }
 
   handleItemSelect = (itemId, _, time) => {
-    console.log('Selected: ' + itemId, moment(time).format())
+    console.log('Selected: ' + itemId, formatISO(time))
   }
 
   handleItemDoubleClick = (itemId, _, time) => {
-    console.log('Double Click: ' + itemId, moment(time).format())
+    console.log('Double Click: ' + itemId, formatISO(time))
   }
 
   handleItemContextMenu = (itemId, _, time) => {
-    console.log('Context Menu: ' + itemId, moment(time).format())
+    console.log('Context Menu: ' + itemId, formatISO(time))
   }
 
   handleItemMove = (itemId, dragTime, newGroupOrder) => {