+++ demo/app/demo-main/index.js
@@ -1,26 +1,18 @@
 /* eslint-disable no-console */
 import React, { Component } from 'react'
-import moment from 'moment'
 
 import Timeline, {
   TimelineMarkers,
-  TimelineHeaders,
   TodayMarker,
   CustomMarker,
   CursorMarker,
-  CustomHeader,
-  SidebarHeader,
-  DateHeader
 } from 'react-calendar-timeline'
 
 import generateFakeData from '../generate-fake-data'
+import { add, formatISO, startOfDay } from 'date-fns'
 
-var minTime = moment()
-  .add(-6, 'months')
-  .valueOf()
-var maxTime = moment()
-  .add(6, 'months')
-  .valueOf()
+var minTime = add(new Date(), {months: -6}).valueOf()
+var maxTime = add(new Date(), {months: 6}).valueOf()
 
 var keys = {
   groupIdKey: 'id',
@@ -39,13 +31,8 @@ export default class App extends Component {
     super(props)
 
     const { groups, items } = generateFakeData()
-    const defaultTimeStart = moment()
-      .startOf('day')
-      .toDate()
-    const defaultTimeEnd = moment()
-      .startOf('day')
-      .add(1, 'day')
-      .toDate()
+    const defaultTimeStart = startOfDay(new Date())
+    const defaultTimeEnd = add(startOfDay(new Date()), {days: 1})
 
     this.state = {
       groups,
@@ -56,31 +43,31 @@ export default class App extends Component {
   }
 
   handleCanvasClick = (groupId, time) => {
-    console.log('Canvas clicked', groupId, moment(time).format())
+    console.log('Canvas clicked', groupId, formatISO(time))
   }
 
   handleCanvasDoubleClick = (groupId, time) => {
-    console.log('Canvas double clicked', groupId, moment(time).format())
+    console.log('Canvas double clicked', groupId, formatISO(time))
   }
 
   handleCanvasContextMenu = (group, time) => {
-    console.log('Canvas context menu', group, moment(time).format())
+    console.log('Canvas context menu', group, formatISO(time))
   }
 
   handleItemClick = (itemId, _, time) => {
-    console.log('Clicked: ' + itemId, moment(time).format())
+    console.log('Clicked: ' + itemId, formatISO(time))
   }
 
   handleItemSelect = (itemId, _, time) => {
-    console.log('Selected: ' + itemId, moment(time).format())
+    console.log('Selected: ' + itemId, formatISO(time))
   }
 
   handleItemDoubleClick = (itemId, _, time) => {
-    console.log('Double Click: ' + itemId, moment(time).format())
+    console.log('Double Click: ' + itemId, formatISO(time))
   }
 
   handleItemContextMenu = (itemId, _, time) => {
-    console.log('Context Menu: ' + itemId, moment(time).format())
+    console.log('Context Menu: ' + itemId, formatISO(time))
   }
 
   handleItemMove = (itemId, dragTime, newGroupOrder) => {
@@ -135,6 +122,10 @@ export default class App extends Component {
     }
   }
 
+  handleZoom = (timelineContext, unit) => {
+    console.log('Zoomed', timelineContext, unit)
+  }
+
   moveResizeValidator = (action, item, time) => {
     if (time < new Date().getTime()) {
       var newTime =
@@ -174,22 +165,21 @@ export default class App extends Component {
         onItemResize={this.handleItemResize}
         onItemDoubleClick={this.handleItemDoubleClick}
         onTimeChange={this.handleTimeChange}
+        onZoom={this.handleZoom}
         moveResizeValidator={this.moveResizeValidator}
+        buffer={3}
       >
         <TimelineMarkers>
           <TodayMarker />
           <CustomMarker
             date={
-              moment()
-                .startOf('day')
+              startOfDay(new Date())
                 .valueOf() +
               1000 * 60 * 60 * 2
             }
           />
           <CustomMarker
-            date={moment()
-              .add(3, 'day')
-              .valueOf()}
+            date={add(new Date(), {days: 3}).valueOf()}
           >
             {({ styles }) => {
               const newStyles = { ...styles, backgroundColor: 'blue' }