+++ demo/app/demo-vertical-classes/index.js
@@ -1,16 +1,16 @@
-import moment from "moment";
 import React, {Component} from "react";
 
 import generateFakeData from '../generate-fake-data'
 import Timeline from "react-calendar-timeline";
+import { add, getHours, getMinutes, isSameDay, parse, startOfDay } from 'date-fns'
 
-const format = "DD.MM.YYYY"
-const holidays = [moment("01.01.2018", format), moment("06.01.2018", format), moment("30.03.2018", format),
-  moment("01.04.2018", format), moment("02.04.2018", format), moment("01.05.2018", format),
-  moment("10.05.2018", format), moment("20.05.2018", format), moment("21.05.2018", format),
-  moment("31.05.2018", format), moment("15.08.2018", format), moment("26.10.2018", format),
-  moment("01.11.2018", format), moment("08.12.2018", format), moment("24.12.2018", format),
-  moment("25.12.2018", format), moment("26.12.2018", format), moment("31.12.2018", format)]
+const format = "dd.MM.yyyy"
+const holidays = [parse("01.01.2018", format, new Date()), parse("06.01.2018", format, new Date()), parse("30.03.2018", format, new Date()),
+  parse("01.04.2018", format, new Date()), parse("02.04.2018", format, new Date()), parse("01.05.2018", format, new Date()),
+  parse("10.05.2018", format, new Date()), parse("20.05.2018", format, new Date()), parse("21.05.2018", format, new Date()),
+  parse("31.05.2018", format, new Date()), parse("15.08.2018", format, new Date()), parse("26.10.2018", format, new Date()),
+  parse("01.11.2018", format, new Date()), parse("08.12.2018", format, new Date()), parse("24.12.2018", format, new Date()),
+  parse("25.12.2018", format, new Date()), parse("26.12.2018", format, new Date()), parse("31.12.2018", format, new Date())]
 
 const keys = {
   groupIdKey: 'id',
@@ -30,13 +30,8 @@ export default class App extends Component {
     super(props)
 
     const {groups, items} = generateFakeData()
-    const defaultTimeStart = moment()
-      .startOf('day')
-      .toDate()
-    const defaultTimeEnd = moment()
-      .startOf('day')
-      .add(1, 'day')
-      .toDate()
+    const defaultTimeStart = startOfDay(new Date())
+    const defaultTimeEnd = add(startOfDay(new Date()), {days: 1})
 
     this.state = {
       groups,
@@ -47,25 +42,25 @@ export default class App extends Component {
   }
 
   getMinutesOfDay = (date) => {
-    return date.hours() * 60 + date.minutes()
+    return getHours(date) * 60 + getMinutes(date)
   }
 
   verticalLineClassNamesForTime = (timeStart, timeEnd) => {
-    const currentTimeStart = moment(timeStart)
-    const currentTimeEnd = moment(timeEnd)
+    const currentTimeStart = timeStart
+    const currentTimeEnd = timeEnd
 
     let classes = [];
 
     // check for public holidays
     for (let holiday of holidays) {
-      if (holiday.isSame(currentTimeStart, "day") && holiday.isSame(currentTimeEnd, "day")) {
+      if (isSameDay(holiday, currentTimeStart) && isSameDay(holiday, currentTimeEnd)) {
         classes.push("holiday")
       }
     }
 
     // highlight lunch break (12:00-13:00)
-    const lunchStart = moment().hours(12).minutes(0).seconds(0);
-    const lunchEnd = moment().hours(13).minutes(0).seconds(0);
+    const lunchStart = new Date().setHours(12, 0, 0);
+    const lunchEnd = new Date().setHours(13, 0, 0);
     if (this.getMinutesOfDay(currentTimeStart) >= this.getMinutesOfDay(lunchStart) &&
       this.getMinutesOfDay(currentTimeEnd) <= this.getMinutesOfDay(lunchEnd)) {
       classes.push("lunch");